* commit '35a8f63cf5f3785d38c0ef0a27567e10d5e7dae1':
  [EDELIVERY-6601] addressing PR comments: add comment
  [EDELIVERY-6601] updated changelog.txt and upgrade-info.txt
  [EDELIVERY-6601] reverted back
  [EDELIVERY-6601] reverted back the wrongly commited files
  [EDELIVERY-6601] addressing PR comments: reverted back to maxSize="2147483647"
  [EDELIVERY-6601] correction
  [EDELIVERY-6601] added flag to mark aync saved payloads
  [EDELIVERY-6601] upgraded maxSize attribute to long type replaced maxSize value with the new maximal one - 9223372036854775807 added one more IT test added update and full script changes for Liquibase
  [EDELIVERY-6601] started to update maxSize to Long
  [EDELIVERY-6601] implemented async saving maxSize validation for payloads

+++ Domibus-MSH/src/main/java/eu/domibus/core/pmode/validation/validators/PayloadProfilesValidator.java
@@ -55,10 +55,10 @@ public class PayloadProfilesValidator implements PModeValidator {
         }
 
         //validate max Size
-        int maxSize = payloadProfile.getMaxSize();
+        long maxSize = payloadProfile.getMaxSize();
         if (maxSize < 0) {
             issues.add(createIssue(payloadProfile, String.valueOf(maxSize),
-                    "the maxSize value [%s] of payload profile [%s] should be neither negative neither a positive value greater than " + Integer.MAX_VALUE));
+                    "the maxSize value [%s] of payload profile [%s] should be neither negative neither a positive value greater than " + Long.MAX_VALUE));
         }
 
         return issues;
